server/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portscanner-server"
version = "1.0.0"
description = "Port Scanner Delta Reporter – Server"
readme = "README.md"
authors = [{name = "Jared Stanbrook", email = "jared@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "alembic==1.17.1",
    "APScheduler==3.11.0",
    "Authlib==1.6.5",
    "bidict==0.23.1",
    "blinker==1.9.0",
    "cachelib==0.13.0",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cryptography==46.0.3",
    "Flask==3.1.2",
    "Flask-Login==0.6.3",
    "Flask-Migrate==4.1.0",
    "Flask-Session==0.8.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-WTF==1.2.2",
    "h11==0.16.0",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "Mako==1.3.10",
    "MarkupSafe==3.0.3",
    "msal==1.34.0",
    "msgspec==0.19.0",
    "numpy==2.3.4",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==12.0.0",
    "psutil==7.1.2",
    "pycparser==2.23",
    "PyJWT==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-engineio==4.12.3",
    "python-nmap==0.7.1",
    "pytz==2025.2",
    "reportlab==4.4.4",
    "requests==2.32.5",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "SQLAlchemy==2.0.44",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "Werkzeug==3.1.3",
    "wsproto==1.2.0",
    "WTForms==3.2.1",
    "gunicorn==23.0.0",
    "gevent==25.9.1",
]

[project.optional-dependencies]
dev = [
    "black==23.7.0",
    "flake8==6.0.0",
    "mccabe==0.7.0",
    "mypy_extensions==1.1.0",
    "pathspec==0.12.1",
    "platformdirs==4.4.0",
    "pluggy==1.6.0",
    "pycodestyle==2.10.0",
    "pyflakes==3.0.1",
    "pytest==7.4.2",
    "pytest-flask==1.2.0",
    "iniconfig==2.1.0",
]

[project.scripts]
portscanner-server = "run:main"

[tool.setuptools]
py-modules = ["run"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]